pyproject.toml
[tool.poetry]
name = "looqbox"
version = "0.1.0"
description = ""
authors = ["Herbert Souto <herbert-93@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mysql-connector-python = "^9.1.0"
pandas = "^2.2.3"
matplotlib = "^3.9.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
